+++ packages/extension-koni-base/src/api/dotsama/transfer.ts
@@ -100,7 +100,7 @@ export async function checkSupportTransfer (networkKey: string, token: string, d
   } else if (isTxBalancesSupported && (!tokenInfo || tokenInfo.isMainToken)) {
     result.supportTransfer = true;
     result.supportTransferAll = true;
-  } else if (['pioneer'].includes(networkKey) && tokenInfo && tokenInfo.symbol === 'BIT') {
+  } else if (['pioneer', 'bitcountry'].includes(networkKey) && tokenInfo && tokenInfo.symbol === 'BIT') {
     result.supportTransfer = true;
     result.supportTransferAll = true;
   } else if (['statemint', 'statemine'].includes(networkKey) && tokenInfo) {
@@ -181,7 +181,7 @@ export async function estimateFee (
 
       fee = paymentInfo.partialFee.toString();
     }
-  } else if (['pioneer'].includes(networkKey) && tokenInfo && tokenInfo.symbol === 'BIT') {
+  } else if (['pioneer', 'bitcountry'].includes(networkKey) && tokenInfo && tokenInfo.symbol === 'BIT') {
     const paymentInfo = await api.tx.currencies.transfer(to, tokenInfo.specialOption, value).paymentInfo(fromKeypair);
 
     fee = paymentInfo.partialFee.toString();
@@ -261,7 +261,7 @@ export function updateTransferResponseTxResult (
         response.txResult.change = record.event.data[3]?.toString() || '0';
         response.txResult.changeSymbol = tokenInfo.symbol;
       }
-    } else if (['pioneer', 'bit.country'].includes(networkKey) && tokenInfo && !tokenInfo.isMainToken) {
+    } else if (['pioneer', 'bitcountry'].includes(networkKey) && tokenInfo && !tokenInfo.isMainToken) {
       if (record.event.section === 'tokens' &&
         record.event.method.toLowerCase() === 'transfer') {
         response.txResult.change = record.event.data[3]?.toString() || '0';
@@ -472,7 +472,7 @@ export const createTransferExtrinsic = async ({ apiProp, from, networkKey, to, t
     } else if (value) {
       transfer = api.tx.kton.transfer(to, new BN(value));
     }
-  } else if (['pioneer'].includes(networkKey) && tokenInfo && tokenInfo.symbol === 'BIT') {
+  } else if (['pioneer', 'bitcountry'].includes(networkKey) && tokenInfo && tokenInfo.symbol === 'BIT') {
     transfer = api.tx.currencies.transfer(to, tokenInfo.specialOption, value);
   } else if (['statemint', 'statemine'].includes(networkKey) && tokenInfo && !tokenInfo.isMainToken) {
     transfer = api.tx.assets.transfer(tokenInfo.assetIndex, to, value);